requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
pyarrow==17.0.0
plotly==5.24.1
streamlit==1.38.0
python-dotenv==1.0.1
folium==0.17.0
pydeck==0.9.1
pytest==8.3.3
ruff==0.6.8
black==24.8.0
joblib==1.4.2
